requirements.txt
Flask
flask-cors
pdf2image
python-pptx
python-pptx[dev]